+++ spec/images.spec.js
@@ -160,7 +160,7 @@ describe('Images addon', function () {
         });
     });
 
-    it('truggers input event twice on showImage for preview', function (done) {
+    it('triggers input event twice on showImage for preview', function (done) {
         var inputTriggerCount = 0,
             stubbedImage = jasmine.createSpy('image'),
             context = this.$el.prepend('<div class="medium-insert-images medium-insert-active">' +
@@ -182,6 +182,38 @@ describe('Images addon', function () {
             context: context
         });
         stubbedImage.onload();
+    });
+
+    it('calls uploadCompleted when preview is enabled', function (done) {
+        var stubbedImage = jasmine.createSpy('image'),
+            context = this.$el.prepend('<div class="medium-insert-images medium-insert-active">' +
+                '<figure><img src="data:" alt=""></figure>' +
+            '</div>');
+
+        spyOn(this.addon, 'getDOMImage').and.returnValue(stubbedImage);
+
+        this.addon.options.uploadCompleted = function () {
+            expect(true).toBe(true);
+            done();
+        };
+
+        this.addon.showImage('http://image.co', {
+            context: context
+        });
+        stubbedImage.onload();
+    });
+
+    it('calls uploadCompleted when preview is disabled', function (done) {
+        this.addon.options.preview = false;
+
+        this.addon.options.uploadCompleted = function () {
+            expect(true).toBe(true);
+            done();
+        };
+
+        this.addon.showImage(null, {
+            submit: function () {}
+        });
     });
 
     it('supports selecting image', function () {
@@ -349,21 +381,23 @@ describe('Images addon', function () {
         $('.medium-insert-images-toolbar .medium-editor-action').first().click();
     });
 
-    it('calls uploadCompleted callback', function (done) {
-        this.addon.options.uploadCompleted = function () {
+    it('calls callback function after clicking on image action ', function (done) {
+        this.addon.options.actions.remove.clicked = function () {
             expect(true).toBe(true);
             done();
         };
 
-        spyOn(this.addon, 'showImage');
+        this.$el.find('p')
+            .addClass('medium-insert-images')
+            .append('<figure><img src="image1.jpg" alt=""></figure>');
 
-        this.addon.uploadDone(null, {
-            result: {
-                files: [
-                    { url: 'test.jpg' }
-                ]
-            }
-        });
+        // Place caret into first paragraph
+        placeCaret(this.$el.find('p').get(0), 0);
+
+        this.$el.find('img').click();
+        jasmine.clock().tick(50);
+
+        $('.medium-insert-images-toolbar2 .medium-editor-action').first().click();
     });
 
     it('validatates file type on upload', function (done) {